requirements.txt
asgiref==3.5.2
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
psycopg2 @ file:///opt/conda/conda-bld/psycopg2_1662147651466/work
pyparsing==3.0.9
pytz==2022.5
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
